+++ src/Filesystem.php
@@ -10,6 +10,7 @@ use Innmind\Filesystem\{
 use Innmind\Url\Path;
 use Innmind\FileWatch\Ping;
 use Innmind\Immutable\{
+    Attempt,
     Maybe,
     Str,
     Sequence,
@@ -36,7 +37,7 @@ interface Filesystem
      *
      * @param Sequence<Maybe<Str>> $chunks
      *
-     * @return Maybe<Content>
+     * @return Attempt<Content>
      */
-    public function temporary(Sequence $chunks): Maybe;
+    public function temporary(Sequence $chunks): Attempt;
 }